rename stereotype dir singlar to plural
