+++ README.md
@@ -17,4 +17,4 @@
 
 * [extensible-custom-error](https://github.com/necojackarc/extensible-custom-error)
   * ライセンス： https://github.com/necojackarc/extensible-custom-error/blob/52d56448d9f535835a9ffbc7e447b951555c08c2/README.md
-  * 利用箇所： [ApplicationError.ts](example-app/SantokuApp/src/bases/core/error/ApplicationError.ts)
+  * 利用箇所： [ApplicationError.ts](example-app/SantokuApp/src/bases/core/errors/ApplicationError.ts)
