+++ example-app/SantokuApp/src/apps/app/errors/initialDataError.ts
@@ -1,4 +1,4 @@
-import {ApplicationError} from 'bases/core/error/ApplicationError';
+import {ApplicationError} from 'bases/core/errors/ApplicationError';
 
 export class InitialDataError extends ApplicationError {}
