+++ example-app/SantokuApp/src/bases/core/configs/AppConfig.test.ts
