+++ example-app/SantokuApp/src/bases/core/configs/AppConfig.ts
