+++ example-app/SantokuApp/src/bases/core/errors/ApplicationError.ts
