+++ example-app/SantokuApp/src/features/account/errors/ActiveAccountIdNotFoundError.ts
@@ -1,4 +1,4 @@
-import {ApplicationError} from 'bases/core/error/ApplicationError';
+import {ApplicationError} from 'bases/core/errors/ApplicationError';
 
 /** アクティブなアカウントIDがセキュアストレージに存在しない場合に送出するエラー */
 export class ActiveAccountIdNotFoundError extends ApplicationError {}
