+++ example-app/SantokuApp/src/features/account/errors/PasswordNotFoundError.ts
@@ -1,4 +1,4 @@
-import {ApplicationError} from 'bases/core/error/ApplicationError';
+import {ApplicationError} from 'bases/core/errors/ApplicationError';
 
 /** パスワードがセキュアストレージに存在しない場合に送出するエラー */
 export class PasswordNotFoundError extends ApplicationError {}
