+++ example-app/SantokuApp/src/features/account/errors/UnauthorizedError.ts
@@ -1,4 +1,4 @@
-import {ApplicationError} from 'bases/core/error/ApplicationError';
+import {ApplicationError} from 'bases/core/errors/ApplicationError';
 
 export class UnauthorizedError extends ApplicationError {}
