+++ example-app/SantokuApp/src/features/app-updates/services/openStoreLink.ts
@@ -1,4 +1,4 @@
-import {AppConfig} from 'bases/core/config/AppConfig';
+import {AppConfig} from 'bases/core/configs/AppConfig';
 import {log} from 'bases/logging';
 import {Linking} from 'react-native';
