+++ example-app/SantokuApp/src/features/backend/utils/customInstance.ts
@@ -1,7 +1,7 @@
 import Axios, {AxiosError, AxiosInstance, AxiosRequestConfig, AxiosResponse} from 'axios';
-import {AppConfig} from 'bases/core/config/AppConfig';
+import {AppConfig} from 'bases/core/configs/AppConfig';
 import {applicationName, nativeApplicationVersion} from 'expo-application';
-import {RequestTimeoutError} from 'features/backend/error/RequestTimeoutError';
+import {RequestTimeoutError} from 'features/backend/errors/RequestTimeoutError';
 import {Platform} from 'react-native';
 
 export type ErrorType<Error> = AxiosError<Error>;
