+++ example-app/SantokuApp/src/features/demo-config/constants/info.ts
@@ -1,4 +1,4 @@
-import {AppConfig} from 'bases/core/config/AppConfig';
+import {AppConfig} from 'bases/core/configs/AppConfig';
 
 export const info = [
   {
