+++ example-app/SantokuApp/src/features/demo-http-api/services/fetchWithAxios.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import {AppConfig} from 'bases/core/config/AppConfig';
+import {AppConfig} from 'bases/core/configs/AppConfig';
 
 export const fetchWithAxios = (withCredentials: boolean, maxRedirects?: number) => {
   return axios.get<{message: string}>(AppConfig.santokuAppBackendUrl + '/api/sandbox/fetch_test/redirect', {
