+++ example-app/SantokuApp/src/features/demo-push-notification/services/notifyMessageToAll.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import {AppConfig} from 'bases/core/config/AppConfig';
+import {AppConfig} from 'bases/core/configs/AppConfig';
 
 export const notifyMessageToAll = (channelId?: string) =>
   axios.put(`${AppConfig.santokuAppBackendUrl}/api/sandbox/push-notification/all`, {
