+++ example-app/SantokuApp/src/features/demo-push-notification/services/notifyMessageToMe.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import {AppConfig} from 'bases/core/config/AppConfig';
+import {AppConfig} from 'bases/core/configs/AppConfig';
 
 export const notifyMessageToMe = (token: string, channelId?: string) =>
   axios.put(`${AppConfig.santokuAppBackendUrl}/api/sandbox/push-notification/single/${token}`, {
