+++ example-app/SantokuApp/src/fixtures/msw/utils/backendUrl.ts
@@ -1,3 +1,3 @@
-import {AppConfig} from 'bases/core/config/AppConfig';
+import {AppConfig} from 'bases/core/configs/AppConfig';
 
 export const backendUrl = `${AppConfig.santokuAppBackendUrl}/api`;
